customer resource created

+++ routes/web.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\ProductController;
 use App\Http\Controllers\RegisterController;
 use App\Http\Controllers\OrderController;
 use App\Http\Controllers\ProductPurchaseController;
+use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
 use App\Models\Product;
 use App\Models\CartItem;
@@ -41,10 +42,10 @@ Route::get('/dashboard', function () {
         }
     }
     return view('dashboard')->with("products", $products)->with("cartItem", $cartItem ?? null)->with('count', $count ?? null )->with('total',$total ?? null);
-})->middleware("role")->name("dashboard");
+})->middleware("user")->name("dashboard");
 
-Route::get('/product/{product}', [ProductPurchaseController::class,'create'])->middleware("role")->name("product.purchase");
-Route::post('/product/store', [ProductPurchaseController::class,'store'])->middleware("role")->name("product.store");
+Route::get('/product/{product}', [ProductPurchaseController::class,'create'])->middleware("user")->name("product.purchase");
+Route::post('/product/store', [ProductPurchaseController::class,'store'])->middleware("user")->name("product.store");
 
 Route::get('/confirm_purchase/{product}', function (Product $product) {
     $user = auth()->user();
@@ -61,7 +62,7 @@ Route::get('/confirm_purchase/{product}', function (Product $product) {
         }
     }
     return view('confirm_purchase')->with('product',$product)->with("cartItem", $cartItem ?? null)->with('count',$count ?? null)->with('total',$total ?? null);
-})->name("confirm.purchase")->middleware("auth");
+})->middleware("user")->name("confirm.purchase");
 
 Route::get('/order_purchase', function(){
     
@@ -78,12 +79,12 @@ Route::get('/order_purchase', function(){
         }
     }
     return view('order_purchase')->with("cartItem", $cartItem ?? null )->with('total',$total ?? null);
-})->middleware("auth")->name("order.purchase");
+})->middleware("user")->name("order.purchase");
 
-Route::get('/confirm_order', [OrderController::class,'create'])->name('confirm.order');
-Route::post('/confirm_order/store', [OrderController::class,'store'])->name('confirm.order.store');
-Route::get('/my-orders', [OrderController::class,'index'])->name('my_orders');
-Route::get('/my-orders/{order_id}', [OrderController::class,'show'])->name('my_order');
+Route::get('/confirm_order', [OrderController::class,'create'])->middleware("user")->name('confirm.order');
+Route::post('/confirm_order/store', [OrderController::class,'store'])->middleware("user")->name('confirm.order.store');
+Route::get('/my-orders', [OrderController::class,'index'])->middleware("user")->name('my_orders');
+Route::get('/my-orders/{order_id}', [OrderController::class,'show'])->middleware("user")->name('my_order');
 
 Route::get('/about', function () {
     return view('about_us');
@@ -99,18 +100,21 @@ Route::get('/product', function ( ) {
 
 Route::get('/admin/login', [AdminController::class,'create'])->name('admin.login');
 Route::post('/admin/store', [AdminController::class,'store'])->name('admin.store');
-Route::get('/admin/logout', [AdminController::class,'destroy'])->name('admin.logout');
+Route::get('/admin/logout', [AdminController::class,'destroy'])->middleware("admin")->name('admin.logout');
 
-Route::get('/all_products', [ProductController::class,'index'])->name('all_products');
-Route::get('/new_product', [ProductController::class,'create'])->name('new_product');
-Route::post('/new_product/store', [ProductController::class,'store'])->name('new_product.store');
-Route::get('/edit_product/{product}', [ProductController::class,'edit'])->name('edit_product');
-Route::post('/update_product/{product}', [ProductController::class,'update'])->name('update_product');
-Route::delete('/delete_product/{product}', [ProductController::class,'destroy'])->name('delete_product');
+Route::get('/all_products', [ProductController::class,'index'])->middleware("admin")->name('all_products');
+Route::get('/new_product', [ProductController::class,'create'])->middleware("admin")->name('new_product');
+Route::post('/new_product/store', [ProductController::class,'store'])->middleware("admin")->name('new_product.store');
+Route::get('/edit_product/{product}', [ProductController::class,'edit'])->middleware("admin")->name('edit_product');
+Route::post('/update_product/{product}', [ProductController::class,'update'])->middleware("admin")->name('update_product');
+Route::delete('/delete_product/{product}', [ProductController::class,'destroy'])->middleware("admin")->name('delete_product');
+
+Route::get('/all_users', [UserController::class,'index'])->middleware("admin")->name('all_users');
+Route::delete('/delete_user/{user}', [UserController::class,'destroy'])->middleware("admin")->name('delete_user');
 
 Route::get('/admin/dashboard', function (){
     return view('admin.dashboard');
-})->name("admin.dashboard");
+})->middleware("admin")->name("admin.dashboard");
